PVP-229 Separate calendar logic into different files, parts; Sort tasks by various factors.
